Move FrequencyCalculator to commons
