+++ ihmc-avatar-interfaces/src/main/java/us/ihmc/avatar/colorVision/stereo/DualBlackflyUDPReceiver.java
@@ -5,7 +5,7 @@ import us.ihmc.log.LogTools;
 import us.ihmc.perception.ImageDimensions;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.robotics.robotSide.SideDependentList;
-import us.ihmc.tools.time.FrequencyCalculator;
+import us.ihmc.commons.time.FrequencyCalculator;
 
 import java.io.IOException;
 import java.net.DatagramPacket;
