+++ ihmc-graphics/src/libgdx/java/us/ihmc/rdx/imgui/ImPlotFrequencyPlot.java
@@ -1,6 +1,6 @@
 package us.ihmc.rdx.imgui;
 
-import us.ihmc.tools.time.FrequencyCalculator;
+import us.ihmc.commons.time.FrequencyCalculator;
 
 public class ImPlotFrequencyPlot
 {
