+++ ihmc-graphics/src/libgdx/java/us/ihmc/rdx/vr/RDXVRManager.java
@@ -19,7 +19,7 @@ import us.ihmc.rdx.sceneManager.RDX3DScene;
 import us.ihmc.rdx.ui.RDXBaseUI;
 import us.ihmc.rdx.ui.gizmo.RDXPose3DGizmo;
 import us.ihmc.robotics.robotSide.RobotSide;
-import us.ihmc.tools.time.FrequencyCalculator;
+import us.ihmc.commons.time.FrequencyCalculator;
 
 import java.util.ArrayList;
 import java.util.List;
